# wa-gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: WA Gateway API
  description: WhatsApp API Gateway built on Baileys — self-hosted alternative to Green API
  version: 1.0.0
  contact:
    name: WA Gateway
servers:
  - url: /v1
    description: API v1
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-Key
      in: header
      description: Instance API key (returned on creation) or Master API key
  schemas:
    Instance:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [STARTING, QR_READY, CONNECTED, DISCONNECTED] }
        phoneNumber: { type: string, nullable: true }
        webhookUrl: { type: string, nullable: true }
        settings: { type: object }
        createdAt: { type: string, format: date-time }
...
paths:
  /instances:
    post:
      summary: Create new instance
      operationId: createInstance
      tags: [Instances]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                webhookUrl: { type: string, format: uri }
                settings:
                  type: object
                  properties:
                    delaySend: { type: number }
                    keepOnline: { type: boolean }
      responses:
        '201':
          description: Instance created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Instance'
                  - type: object
                    properties:
                      apiKey: { type: string }
        '400': { description: Invalid webhook URL }

    get:
      summary: List instances
      operationId: listInstances
      tags: [Instances]
      responses:
        '200':
          description: Array of instances
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Instance' }
  /instances/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }

    get:
      summary: Get instance details
      operationId: getInstance
      tags: [Instances]
      responses:
        '200': { description: Instance details }
        '404': { description: Not found }

    delete:
      summary: Delete instance
      operationId: deleteInstance
      tags: [Instances]
      responses:
        '200': { description: Deleted }
  /instances/{id}/qr:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get QR code (base64)
      operationId: getQR
      tags: [Instances]
      responses:
        '200':
          description: QR code image
          content:
            application/json:
              schema:
                type: object
                properties:
                  type: { type: string }
                  image: { type: string, description: base64 data URI }
        '409': { description: QR not ready }
  /instances/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get connection status
      operationId: getStatus
      tags: [Instances]
      responses:
        '200':
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  phoneNumber: { type: string, nullable: true }
  /instances/{id}/logout:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Logout (keep instance)
      operationId: logoutInstance
      tags: [Instances]
      responses:
        '200': { description: Logged out }
  /instances/{id}/settings:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      summary: Update instance settings
      operationId: updateSettings
      tags: [Instances]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                webhookUrl: { type: string, format: uri }
                delaySend: { type: number }
                keepOnline: { type: boolean }
      responses:
        '200': { description: Updated }
...